Created a HealthContext to allow for querying by other methods at anytime
